Change membership confirmation deadline
The original process was intended to take place
at the end of 2021, however we decided to hold off
until 2022. As a result, by changing the deadline to
the end of the following calendar month, whenever the
email gets sent out, there will be a reasonable amount
of time between delivery and termination of membership.